dynamic-threads: resync with trunk
git-svn-id: http://streamboard.de.vu/svn/oscam/branches/dynamic-threads@5593 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ global-functions.h
@@ -424,4 +424,4 @@ extern int32_t reader_emm(struct s_reader * reader, EMM_PACKET *);
 extern int32_t reader_get_emm_type(EMM_PACKET *ep, struct s_reader * reader);
 extern struct s_cardsystem *get_cardsystem_by_caid(uint16_t caid);
 extern void reader_device_close(struct s_reader * reader);
-extern int8_t cs_emmlen_is_blocked(struct s_reader *rdr, int8_t len);
+extern int8_t cs_emmlen_is_blocked(struct s_reader *rdr, int16_t len);
